pyproject.toml
[project]
name = "docu-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.46",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.0",
]
